requirements.txt
streamlit==1.37.1
pandas==2.3.2
numpy==2.0.1
requests==2.32.3
python-dateutil==2.9.0
pytz==2024.1
